water_Notifications/ephemeral_test/requirements.txt

amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
bcrypt==3.2.0
billiard==4.2.1
gunicorn==23.0.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
databases==0.9.0
dnspython==2.7.0
docopt==0.6.2
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-limiter==0.1.6
h11==0.14.0
idna==3.10
celery>=5.4.0
kombu==5.4.2
mysql-connector-python==9.1.0
passlib==1.7.4
pipreqs==0.4.13
prompt_toolkit==3.0.50
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydantic-settings==2.7.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
pytz==2024.2
redis==5.2.1
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
yarg==0.1.10
boto3==1.26.135
botocore==1.29.135
